1、修复了lua5.3中HOOK lua_callk函数处理bug 2、修改了部分代码用来支持skynet框架 3、添加了Babelua对于文件相对路径的处理

+++ src/LuaInject/LuaTypes.h
@@ -55,7 +55,7 @@ extern "C"
 	// =====================================================
 
 	typedef double lua_Number;
-	typedef ptrdiff_t lua_Integer;
+	typedef long long lua_Integer;
 
 	// ===========================
 	// same in Lua 5.1 and Lua 5.2
